+++ src/Revision.php
@@ -3,7 +3,7 @@
 namespace Mediawiki\DataModel;
 
 /**
- * Representation of a page version
+ * Representation of a version of content
  */
 class Revision {
 
@@ -11,64 +11,85 @@ class Revision {
 	 * @var int Id of the revision
 	 */
 	protected $id;
+
 	/**
-	 * @var string
+	 * @var int baserev Id of the revision
+	 */
+	protected $baseRevId;
+
+	/**
+	 * @var mixed
 	 */
 	protected $content;
-	/**
-	 * @var string
-	 */
-	protected $user;
-	/**
-	 * @var EditFlags
-	 */
-	protected $flags;
-	/**
-	 * @var string
-	 */
-	protected $timestamp;
 
-	public function __construct( $id, $content, $user, $flags, $timestamp ) {
+	/**
+	 * @var string initial hash of the content
+	 */
+	protected $initialHash;
+
+	/**
+	 * @param mixed $content
+	 * @param int|null $id
+	 * @param int|null $baseRevId
+	 * @param RevisionDetails|null $revisionDetails
+	 */
+	public function __construct( $content, $id = null, $baseRevId = null, $revisionDetails = null ) {
 		$this->content = $content;
-		$this->flags = $flags;
+		$this->initialHash = sha1( serialize( $content ) );
 		$this->id = $id;
-		$this->timestamp = $timestamp;
-		$this->user = $user;
+		$this->baseRevId = $baseRevId;
+		$this->details = $revisionDetails;
 	}
 
 	/**
-	 * @return string
+	 * @return mixed
 	 */
 	public function getContent() {
 		return $this->content;
 	}
 
 	/**
-	 * @return EditFlags
-	 */
-	public function getFlags() {
-		return $this->flags;
-	}
-
-	/**
-	 * @return int
+	 * @return int|null
 	 */
 	public function getId() {
 		return $this->id;
 	}
 
 	/**
-	 * @return string
+	 * @return int|null
 	 */
-	public function getTimestamp() {
-		return $this->timestamp;
+	public function getBaseRevId() {
+		return $this->baseRevId;
 	}
 
 	/**
-	 * @return string
+	 * @return RevisionDetails|null
 	 */
-	public function getUser() {
-		return $this->user;
+	public function getDetails() {
+		return $this->details;
+	}
+
+	/**
+	 * Has the content been changed since object construction
+	 * @return bool
+	 */
+	public function hasChanged() {
+		return $this->initialHash !== sha1( serialize( $this->content ) );
+	}
+
+	/**
+	 * Convenience function for creating new revisions
+	 *
+	 * @param Revision $revision
+	 *
+	 * @return Revision
+	 */
+	public static function newFromRevision( $revision ) {
+		$content = $revision->getContent();
+		if( is_object( $content ) ) {
+			$content = clone $content;
+		}
+		return new self( $content , null, $revision->getId() );
 	}
 
 }