+++ RELEASENOTES.md
@@ -2,7 +2,13 @@ These are the release notes for the [mediawiki-datamodel](README.md).
 
 ## Version 0.2 (under development)
 
-* ...
+####Compatability  changes
+
+* Revision enhanced to allow more flexibility, Constructor and public functions have changed
+
+#### Additions
+
+* Introduced the RevisionDetails class
 
 ## Version 0.1 (2014-02-23)